mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP server for MrDocument - provides AI-accessible document search and context tools"
authors = [
    {name = "Developer"}
]
requires-python = ">=3.10,<4"
dependencies = [
    "mcp[cli] (>=1.9.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
]

[project.scripts]
mcp-server = "mcp_server.server:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
